+++ packages/dnd_kit_jaspr/lib/dnd_kit_jaspr.dart
@@ -21,7 +21,6 @@ library;
 
 export 'package:dnd_kit/dnd_kit.dart';
 
-export 'src/a11y/announcements.dart';
 export 'src/a11y/live_region.dart' show DndLiveRegion;
 export 'src/scope/controller.dart';
 export 'src/scope/scope.dart';