+++ packages/dnd_kit_jaspr/lib/src/scope/scope.dart
@@ -1,6 +1,6 @@
+import 'package:dnd_kit/dnd_kit.dart' show DndAnnouncements;
 import 'package:jaspr/jaspr.dart';
 
-import '../a11y/announcements.dart';
 import 'controller.dart';
 
 /// Provides a [DndController] to a Jaspr subtree.